The view itself: skyline panorama, composed relief with contours, hero candidates
purespatial/panorama.py casts rays in true azimuth from the observer through
the 30 m DEM (0.05 degree steps, 30 m samples, 60 km, curvature and
refraction as gdal_viewshed), takes the skyline as the largest angle per ray,
and renders two half-turns at 16:9: distance bands drawn far to near so
nearer ridges occlude farther ones, peaks rated visible standing on the
skyline, peaks rated hidden drawn inside the ridge that hides them, vertical
labels lifted until they fit. The observer stands on the 1 m crest elevation;
a coarse-grid sample there sees the summit cell as a wall.

purespatial/figures.py now composes one RGB relief instead of stacking
layers: a weighted mix of lights (315 dominant), percentile contrast stretch,
slope darkening, vertical exaggeration 1.6 for lidar, a warm paper grey; the
in-view tint is multiplied over the full shade range so the surface stays
legible inside the colour, with the mask edge feathered. Contours with index
lines and optional labels. hero_map reads the DEM at native resolution inside
a 16:9 window (width_m) so 1 m pixels reach the figure. Dark and sightline
variants remain Style knobs.

Engineer Mountain: figures/panorama.png joins the post as the reconstruction
of the view; the near-field map carries 50 m contours with labels; the hero
is the 1 m surface with the whole 3 km viewshed disc (HERO in params.py; the
tight crop, plain surface, panorama, light and dark maps are written beside
it as hero-*.png for comparison). Three panorama tests, 27 in all.

+++ purespatial/dem.py
@@ -114,7 +114,7 @@ def highest_point(path: str | Path, lon: float, lat: float, radius_m: float) -> 
         to_wgs = Transformer.from_crs(ds.crs, "EPSG:4326", always_xy=True)
         x, y = to_map.transform(lon, lat)
         res = ds.res[0]
-        r = int(math.ceil(radius_m / res))
+        r = math.ceil(radius_m / res)
         row, col = ds.index(x, y)
         r0, c0 = max(row - r, 0), max(col - r, 0)
         r1, c1 = min(row + r + 1, ds.height), min(col + r + 1, ds.width)
